+++ lib/QECCore/src/QECCore.jl
@@ -14,10 +14,13 @@ export AbstractECC, AbstractQECC, AbstractCECC, AbstractCSSCode, AbstractDistanc
 export Perfect5, Cleve8, Gottesman
 
 # CSS Codes
-export Toric, Bitflip3, Phaseflip3, Shor9, Steane7, Surface, CSS, QuantumReedMuller, Triangular488, Triangular666
+export Toric, Bitflip3, Phaseflip3, Shor9, Steane7, Surface, CSS, QuantumReedMuller, Triangular488, Triangular666, DelfosseReichardt
 
 # Classical Codes
 export RepCode, ReedMuller, RecursiveReedMuller, Golay, Hamming
+
+# utilities
+export search_self_orthogonal_rm_codes
 
 include("interface.jl")
 include("codes/util.jl")
@@ -43,6 +46,9 @@ include("codes/quantum/color_codes.jl")
 include("codes/classical/reedmuller.jl")
 include("codes/classical/recursivereedmuller.jl")
 include("codes/quantum/quantumreedmuller.jl")
+
+# Delfosse-Reichardt Codes
+include("codes/quantum/delfosse_reichardt_code.jl")
 
 function __init__()
     if isdefined(Base.Experimental, :register_error_hint)