+++ test/test_ecc_base.jl
@@ -313,7 +313,8 @@ const code_instance_args = Dict(
     :LPCode => (c -> (c.A, c.B)).(vcat(LP04, LP118, test_gb_codes, test_bb_codes, test_mbb_codes, test_coprimeBB_codes, test_hcubic_codes, test_twobga_codes, test_honeycomb_color_codes, test_nonabelian_codes, other_lifted_product_codes)),
     :QuantumReedMuller => [3, 4, 5],
     :Triangular488 => [3, 5, 7, 9, 11],
-    :Triangular666 => [3, 5, 7, 9, 11]
+    :Triangular666 => [3, 5, 7, 9, 11],
+    :DelfosseReichardt => [(2,1,3), (2,2,4), (4,3,5), (4,3,6)]
 )
 
 function all_testablable_code_instances(;maxn=nothing)